requirements.txt
# Web Server
Flask==3.0.0
flask-cors==6.0.2
Werkzeug==3.1.6

# AWS
boto3==1.42.68
python-dotenv==1.0.0

# Machine Learning
xgboost==2.0.3
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.2

# Simulation & Network
requests==2.31.0
